refactor(capabilities): restructure the inventory cap to the canonical shape
Brings aether.inventory to the ADR-0122/0123 canonical cap shape,
mirroring the component/ runtime-directory exemplar: mod.rs keeps only
the identity ZST (with `#[actor(singleton)]` lifted onto the struct)
and the always-on kind imports; the `#[runtime] impl NativeActor`, the
InventoryCapabilityState, and the cap's tests move to
inventory/runtime/mod.rs; manifest.rs and resolve.rs nest as
runtime/manifest.rs and runtime/resolve.rs behind the one
`feature = "runtime"` gate. The module header now cites the ADR-0121
harness must-stay rule for the absent kinds.rs (the
Manifest/Resolve/ListKinds/ListHandlers family is MCP-consumed and
stays in aether-kinds). No wire, kind, or behavior change.

Closes #2475.
